+++ src/main/java/com/agenticcp/core/common/enums/CommonErrorCode.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.common.enums;
 
-import com.agenticcp.core.common.dto.BaseErrorCode;
+import com.agenticcp.core.common.dto.exception.BaseErrorCode;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
