+++ src/main/java/com/agenticcp/core/common/enums/ErrorCategory.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.common.enums;
 
-import com.agenticcp.core.common.dto.BaseErrorCode;
+import com.agenticcp.core.common.dto.exception.BaseErrorCode;
 
 /**
  * API 에러 코드의 카테고리를 정의하는 클래스입니다.
