+++ src/main/java/com/agenticcp/core/controller/HealthController.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.controller;
 
-import com.agenticcp.core.common.dto.ApiResponse;
+import com.agenticcp.core.common.dto.exception.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
