+++ src/main/java/com/agenticcp/core/domain/monitoring/controller/MetricsController.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.domain.monitoring.controller;
 
-import com.agenticcp.core.common.dto.ApiResponse;
+import com.agenticcp.core.common.dto.exception.ApiResponse;
 import com.agenticcp.core.common.exception.BusinessException;
 import com.agenticcp.core.common.exception.ResourceNotFoundException;
 import com.agenticcp.core.common.context.TenantContextHolder;
