+++ src/main/java/com/agenticcp/core/domain/monitoring/controller/TenantCollectorConfigController.java
@@ -1,7 +1,7 @@
 package com.agenticcp.core.domain.monitoring.controller;
 
 import com.agenticcp.core.common.context.TenantContextHolder;
-import com.agenticcp.core.common.dto.ApiResponse;
+import com.agenticcp.core.common.dto.exception.ApiResponse;
 import com.agenticcp.core.domain.monitoring.dto.QuotaRequestDto;
 import com.agenticcp.core.domain.monitoring.dto.TenantCollectorConfigDto;
 import com.agenticcp.core.domain.monitoring.enums.CollectorType;
