+++ src/main/java/com/agenticcp/core/domain/monitoring/enums/MonitoringErrorCode.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.domain.monitoring.enums;
 
-import com.agenticcp.core.common.dto.BaseErrorCode;
+import com.agenticcp.core.common.dto.exception.BaseErrorCode;
 import com.agenticcp.core.common.enums.ErrorCategory;
 import org.springframework.http.HttpStatus;
