+++ src/main/java/com/agenticcp/core/domain/security/enums/SecurityErrorCode.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.domain.security.enums;
 
-import com.agenticcp.core.common.dto.BaseErrorCode;
+import com.agenticcp.core.common.dto.exception.BaseErrorCode;
 import com.agenticcp.core.common.enums.ErrorCategory;
 import org.springframework.http.HttpStatus;
