+++ src/main/java/com/agenticcp/core/domain/tenant/controller/TenantAuthorizationController.java
@@ -1,6 +1,6 @@
 package com.agenticcp.core.domain.tenant.controller;
 
-import com.agenticcp.core.common.dto.ApiResponse;
+import com.agenticcp.core.common.dto.exception.ApiResponse;
 import com.agenticcp.core.domain.tenant.service.TenantAuthorizationService;
 import com.agenticcp.core.domain.user.dto.PermissionResponse;
 import com.agenticcp.core.domain.user.dto.RoleResponse;
