+++ src/test/java/com/agenticcp/core/domain/monitoring/controller/MetricsControllerTest.java
@@ -1,7 +1,7 @@
 package com.agenticcp.core.domain.monitoring.controller;
 
 import com.agenticcp.core.common.context.TenantContextHolder;
-import com.agenticcp.core.common.dto.ApiResponse;
+import com.agenticcp.core.common.dto.exception.ApiResponse;
 import com.agenticcp.core.common.exception.BusinessException;
 import com.agenticcp.core.common.exception.ResourceNotFoundException;
 import com.agenticcp.core.domain.monitoring.entity.Metric;
